IOS-00039 Workout Manager IOS-00039C create WorkoutManager
